fix(geometry): let the sync init path reuse the shared compiled wasm module
Adversarial review of the prewarm caught a hole: `processAdaptive` routes files
under the 2 MB sync threshold (and the bridge/export APIs) through
`IfcLiteBridge.init()`, which calls wasm-bindgen `init()` and lets IT fetch and
compile the binary. That path never consulted the worker pool's compiled-module
memo, so it was deduped against a concurrent prewarm only by the HTTP cache.

Browsers are not required to coalesce in-flight requests for the same URL, so a
small file opened while the prewarm was still streaming could mean two
concurrent ~1.3 MB downloads — on exactly the slow links the prewarm exists to
help, and on exactly the file-size class the prewarm was measured against.

Move the memo into `wasm-shared-module.ts` (a single home both paths import) and
have `IfcLiteBridge.init()` reuse an already-started compile via
`getStartedSharedWasmModule`. That accessor joins an in-flight compile but never
starts one, so a consumer with a working init path and no prewarm behaves
exactly as before. Node still hands over bytes it read off disk and is untouched.

This also removes a pre-existing redundant compile: before, a small file opened
after a large one recompiled the same binary the worker pool had already built.

Verified in Chrome on the production build: opening a 220 KB file after the idle
prewarm now produces exactly ONE wasm fetch for the whole page lifetime (it was
two), and the 24.6 MB worker-pool path still yields an identical 2569 meshes.
Full viewer suite 1765 tests / 0 failures, geometry 146/146, pnpm typecheck green.

Known and unchanged: a prewarm whose fetch stalls is now awaited by the sync
path too, where it would previously have started its own. That exposure already
existed for the worker path (processParallel has awaited the shared compile
since #1818) and is bounded by the browser's socket timeout, after which the
failed URL is evicted and the next attempt refetches.
